Replace system allocator with Rust allocator
Memory management becomes easier when all components of the program are
using the same allocator. For this reason, a Lua allocator function is
provided which allocates through Rust's standard allocator (typically
jemalloc) rather than system malloc.

Accessing the allocator through Vec is chosen to avoid requiring nightly
features. This allocator aligns to usize (4 or 8 bytes), which is less
than malloc usually assumes (8 or 16) but should be enough for Lua's
uses.