Remove readline from lua
The cernan use-case for lua is one that does not use readline. On
systems that attempt to statically linked cernan against MUSL libc the
readline dependency breaks the build on account of the difficulty of
building readline in that environment. I could not determine if there
were a cleaner way to disable readline other than but permanently.
That is, maybe this approach is a hack.

I _think_ disabling readline is okay here as the focus of the library
appears to be the use of lua in a purely embedded environment.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>